Fix a race condition in dummypool that results in flaky `test_end_to_end.test_predicate_on_partition`
The fix is applied only to `DummyPool` and makes it not support empty list ventilation.
Other *Pool implementations have the same issue (see https://github.com/uber/petastorm/issues/71) but
due to a different start-up timing, it does not manifest itself (not saying it's ok)...

It seems to be a design issue, but since we will be moving to use a different, standard, pooling
architecture, we just patch DummyPool since impacts tests stability only.